pyproject.toml
[project]
name = "quick-capture"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-generativeai>=0.8.5",
    "pillow>=12.0.0",
    "pyautogui>=0.9.54",
    "pynput>=1.8.1",
    "python-dotenv>=1.1.1",
]
